* Added VMap stuff (not fully tested yet)
* Fix loading only default Locale

+++ Source/DataExtractor/CASC/Structures/IndexEntry.cs
@@ -2,8 +2,8 @@
 {
     public struct IndexEntry
     {
-        public ushort Index { get; set; }
-        public uint Offset  { get; set; }
-        public uint Size    { get; set; }
+        public int Index { get; set; }
+        public int Offset  { get; set; }
+        public int Size    { get; set; }
     }
 }
